+++ packages/headless/src/features/search/search-actions.ts
@@ -318,7 +318,7 @@ export const buildSearchRequest = (
       filterField: state.folding.fields.collection,
       childField: state.folding.fields.parent,
       parentField: state.folding.fields.child,
-      filterFieldRange: state.folding.numberOfFoldedResults,
+      filterFieldRange: state.folding.filterFieldRange,
     }),
   };
 };